requirements.txt
disnake==2.5.2
dgg-bot==1.6.0
python-dateutil==2.8.2